+++ packages/auth/src/Login/LoginWithPassword.tsx
@@ -20,6 +20,7 @@ import {
   useSignUpState,
   useForgotPasswordActions,
 } from '@frontegg/react-hooks/auth';
+import { FReCaptcha } from '../components/FReCaptcha';
 
 const { Formik } = FFormik;
 
@@ -66,7 +67,7 @@ export const LoginWithPassword: FC<LoginWithPasswordProps> = (props) => {
   }, [shouldDisplayPassword]);
 
   const labelButtonProps = (values: any) => ({
-    testId: 'forgot-password-button',
+    'data-test-id': 'forgotPassBtn',
     disabled: loading,
     onClick: () => {
       setForgotPasswordState({ email: values.email });
@@ -93,10 +94,10 @@ export const LoginWithPassword: FC<LoginWithPasswordProps> = (props) => {
     <>
       {signUpMessage}
       <Formik
-        initialValues={{ email: '', password: '' }}
+        initialValues={{ email: '', password: '', recaptchaToken: '' }}
         validationSchema={validateSchema(validationSchema)}
-        onSubmit={async ({ email, password }) => {
-          shouldDisplayPassword ? login({ email, password }) : preLogin({ email });
+        onSubmit={async ({ email, password, recaptchaToken }) => {
+          shouldDisplayPassword ? login({ email, password, recaptchaToken }) : preLogin({ email });
         }}
       >
         {({ values }) => (
@@ -129,6 +130,7 @@ export const LoginWithPassword: FC<LoginWithPasswordProps> = (props) => {
             </FButton>
 
             <ErrorMessage separator error={error} />
+            <FReCaptcha action='login' />
           </FForm>
         )}
       </Formik>