+++ packages/redux-store/src/auth/SecurityPolicyState/interfaces.ts
@@ -3,6 +3,7 @@ import {
   ISaveSecurityPolicyLockout,
   ISaveSecurityPolicyMfa,
   ISecurityPolicy,
+  ISecurityPolicyCaptcha,
   ISecurityPolicyLockout,
   ISecurityPolicyMfa,
 } from '@frontegg/rest-api';
@@ -14,11 +15,13 @@ type PolicyState<T> = WithStatus & {
 export type GlobalPolicyState = PolicyState<ISecurityPolicy>;
 export type MfaPolicyState = PolicyState<ISecurityPolicyMfa>;
 export type LockoutPolicyState = PolicyState<ISecurityPolicyLockout>;
+export type CaptchaPolicyState = PolicyState<ISecurityPolicyCaptcha>;
 
 export interface SecurityPolicyState {
   globalPolicy: GlobalPolicyState;
   mfaPolicy: MfaPolicyState;
   lockoutPolicy: LockoutPolicyState;
+  captchaPolicy: CaptchaPolicyState;
 }
 
 export type SaveSecurityPolicyMfaPayload = WithCallback<ISaveSecurityPolicyMfa, ISecurityPolicyMfa>;