+++ packages/rest-api/src/auth/interfaces.ts
@@ -8,7 +8,7 @@ export type IPostLogin = {
   SAMLResponse: string;
 };
 
-export type ILogin = { email: string; password: string };
+export type ILogin = { email: string; password: string; recaptchaToken?: string };
 
 export type ILoginResponse = IUserProfile & {
   mfaRequired: boolean;
@@ -131,6 +131,7 @@ export interface IVendorConfig {
 export interface ISignUpUser {
   email: string;
   companyName: string;
+  recaptchaToken?: string;
   name?: string;
   password?: string;
   phoneNumber?: string;
@@ -173,7 +174,7 @@ export interface IDeleteApiToken {
 }
 
 export interface IUpdateSamlRoles {
-  roleIds: Array<string>;
+  roleIds: string[];
 }
 
 export type IGetUserById = { userId: string };